impl createSymbolArgTypeEnhancer
